KP-8262 Work around old Flask not restricting it's dependencies enough
When installing Flask 1.0 on a modern machine, we install dependencies
that are no longer backwards compatible with the calls Flask makes. To
fix this, we must restrict the allowed versions for these dependencies.
Ideally Flask would handle its own requirements in a way that avoids
these kinds of problems, but perhaps compatibility issues of previous
decade releases are not their priority.

The problems are:
- Jinja2 no longer having `Markup` and `escape` (see
  [changelog](https://jinja.palletsprojects.com/en/3.1.x/changes/#version-3-1-0))
- too new `markupsafe` no longer having `soft_unicode` (see
  [changelog](https://markupsafe.palletsprojects.com/en/2.1.x/changes/#version-2-1-0)).
- too new `itsdangerous` no longer having `son` (see
  [changelog](https://itsdangerous.palletsprojects.com/en/2.1.x/changes/#version-2-1-0))